图片验证码、用户登录、注销、列表通用功能、分页、模糊查询、排序

+++ utils/jwts/enter.go
@@ -10,7 +10,7 @@ import (
 
 type Claims struct {
 	UserID uint `json:"user_id"`
-	RoleID uint `json:"role_id"`
+	RoleID int8 `json:"role_id"`
 }
 
 type MyClaims struct {
